requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastmcp==2.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-core==1.0.3
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
logger==1.4
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
openai==1.82.0
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
